Disease detection API/requirements.txt
gunicorn==21.2.0
fastapi==0.100.1
uvicorn==0.20.0
numpy==1.24.3
tensorflow-cpu==2.10.0
Keras-Preprocessing==1.1.2
pillow==9.4.0
python-multipart==0.0.6
uvloop
httptools
pydantic==2.1.1
pydantic_core==2.4.0
requests-oauthlib==1.3.0
scipy==1.10.1
starlette==0.27.0
image==1.5.33